backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.0.1
certifi==2024.12.14
click==8.1.7
coverage==7.6.9
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.7
flake8==7.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy==1.13.0
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pycodestyle==2.12.1
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.9.2
pyobjectid==0.1.3
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
setuptools==59.6.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.41.3
tomli==2.2.1
typer==0.15.1
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
wheel==0.45.1
